Extracts ipc code from background.ts adds constants
