+++ frontend/app/src/background.ts
@@ -1,37 +1,11 @@
 import path from 'path';
-import {
-  app,
-  BrowserWindow,
-  dialog,
-  ipcMain,
-  Menu,
-  nativeTheme,
-  protocol,
-  shell
-} from 'electron';
-import { ProgressInfo } from 'electron-builder';
+import { app, BrowserWindow, Menu, protocol } from 'electron';
 import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer';
-import { autoUpdater } from 'electron-updater';
 import windowStateKeeper from 'electron-window-state';
 import { createProtocol } from 'vue-cli-plugin-electron-builder/lib';
-import { startHttp, stopHttp } from '@/electron-main/http';
-import {
-  BackendOptions,
-  IPC_CHECK_FOR_UPDATES,
-  IPC_DARK_MODE,
-  IPC_DOWNLOAD_PROGRESS,
-  IPC_DOWNLOAD_UPDATE,
-  IPC_INSTALL_UPDATE,
-  IPC_OPEN_PATH,
-  IPC_RESTART_BACKEND,
-  IPC_VERSION,
-  SystemVersion
-} from '@/electron-main/ipc';
-import { debugSettings, getUserMenu } from '@/electron-main/menu';
-import { selectPort } from '@/electron-main/port-utils';
-import { assert } from '@/utils/assertions';
+import { ipcSetup } from '@/electron-main/ipc-setup';
+import { getUserMenu } from '@/electron-main/menu';
 import PyHandler from './py-handler';
-import Timeout = NodeJS.Timeout;
 
 const isDevelopment = process.env.NODE_ENV !== 'production';
 
@@ -56,76 +30,7 @@ const onReady = async () => {
     }
   }
 
-  ipcMain.on('GET_DEBUG', event => {
-    event.returnValue = debugSettings;
-  });
-
-  ipcMain.on('SERVER_URL', event => {
-    event.returnValue = pyHandler.serverUrl;
-  });
-
-  ipcMain.on('PREMIUM_USER_LOGGED_IN', (event, args) => {
-    Menu.setApplicationMenu(Menu.buildFromTemplate(getUserMenu(!args)));
-  });
-  ipcMain.on('CLOSE_APP', async () => await closeApp());
-  ipcMain.on('OPEN_URL', (event, args) => {
-    if (!args.startsWith('https://')) {
-      console.error(`Error: Requested to open untrusted URL: ${args} `);
-      return;
-    }
-    shell.openExternal(args);
-  });
-  ipcMain.on('OPEN_DIRECTORY', async (event, args) => {
-    const directory = await select(args, 'openDirectory');
-    event.sender.send('OPEN_DIRECTORY', directory);
-  });
-  let importTimeout: Timeout;
-  ipcMain.on('METAMASK_IMPORT', async (event, _args) => {
-    try {
-      const port = startHttp(
-        addresses => event.sender.send('METAMASK_IMPORT', { addresses }),
-        await selectPort(40000)
-      );
-      await shell.openExternal(`http://localhost:${port}`);
-      if (importTimeout) {
-        clearTimeout(importTimeout);
-      }
-      importTimeout = setTimeout(() => {
-        stopHttp();
-        event.sender.send('METAMASK_IMPORT', { error: 'waiting timeout' });
-      }, 120000);
-    } catch (e) {
-      event.sender.send('METAMASK_IMPORT', { error: e.message });
-    }
-  });
-
-  ipcMain.on(IPC_RESTART_BACKEND, async (event, options: BackendOptions) => {
-    let success = false;
-    try {
-      assert(win);
-      await pyHandler.exitPyProc(true);
-      await pyHandler.createPyProc(win, options.loglevel);
-      success = true;
-    } catch (e) {
-      console.error(e);
-    }
-
-    event.sender.send(IPC_RESTART_BACKEND, success);
-  });
-
-  ipcMain.on(IPC_VERSION, event => {
-    const version: SystemVersion = {
-      os: process.platform,
-      arch: process.arch,
-      osVersion: process.getSystemVersion(),
-      electron: process.versions.electron
-    };
-    event.sender.send(IPC_VERSION, version);
-  });
-
-  ipcMain.on(IPC_OPEN_PATH, (event, path) => shell.openPath(path));
-
-  setupUpdaterInterop();
+  ipcSetup(pyHandler, win!!, closeApp);
   await createWindow();
 };
 
@@ -152,98 +57,6 @@ if (!lock) {
   app.on('will-quit', async e => {
     e.preventDefault();
     await closeApp();
-  });
-}
-
-async function select(
-  title: string,
-  prop: 'openFile' | 'openDirectory'
-): Promise<string | undefined> {
-  const value = await dialog.showOpenDialog({
-    title,
-    properties: [prop]
-  });
-
-  if (value.canceled) {
-    return undefined;
-  }
-  return value.filePaths?.[0];
-}
-
-function setupUpdaterInterop() {
-  autoUpdater.autoDownload = false;
-  autoUpdater.logger = {
-    error: (message?: any) => pyHandler.logToFile(`(error): ${message}`),
-    info: (message?: any) => pyHandler.logToFile(`(info): ${message}`),
-    debug: (message: string) => pyHandler.logToFile(`(debug): ${message}`),
-    warn: (message?: any) => pyHandler.logToFile(`(warn): ${message}`)
-  };
-  ipcMain.on(IPC_CHECK_FOR_UPDATES, async event => {
-    if (isDevelopment) {
-      console.log('Running in development skipping auto-updater check');
-      return;
-    }
-    autoUpdater.once('update-available', () => {
-      event.sender.send(IPC_CHECK_FOR_UPDATES, true);
-    });
-    autoUpdater.once('update-not-available', () => {
-      event.sender.send(IPC_CHECK_FOR_UPDATES, false);
-    });
-    try {
-      await autoUpdater.checkForUpdates();
-    } catch (e) {
-      console.error(e);
-      pyHandler.logToFile(e);
-      event.sender.send(IPC_CHECK_FOR_UPDATES, false);
-    }
-  });
-
-  ipcMain.on(IPC_DOWNLOAD_UPDATE, async event => {
-    const progress = (progress: ProgressInfo) => {
-      event.sender.send(IPC_DOWNLOAD_PROGRESS, progress.percent);
-      win?.setProgressBar(progress.percent);
-    };
-    autoUpdater.on('download-progress', progress);
-    try {
-      await autoUpdater.downloadUpdate();
-      event.sender.send(IPC_DOWNLOAD_UPDATE, true);
-    } catch (e) {
-      pyHandler.logToFile(e);
-      event.sender.send(IPC_DOWNLOAD_UPDATE, false);
-    } finally {
-      autoUpdater.off('download-progress', progress);
-    }
-  });
-
-  ipcMain.on(IPC_INSTALL_UPDATE, async event => {
-    const quit = new Promise<void>((resolve, reject) => {
-      const quitAndInstall = () => {
-        try {
-          autoUpdater.quitAndInstall();
-          resolve();
-        } catch (e) {
-          pyHandler.logToFile(e);
-          reject(e);
-        }
-      };
-      return setTimeout(quitAndInstall, 5000);
-    });
-    try {
-      await quit;
-      event.sender.send(IPC_INSTALL_UPDATE, true);
-    } catch (e) {
-      pyHandler.logToFile(e);
-      event.sender.send(IPC_INSTALL_UPDATE, e);
-    }
-  });
-
-  ipcMain.on(IPC_DARK_MODE, async (event, enabled) => {
-    if (enabled) {
-      nativeTheme.themeSource = 'dark';
-    } else {
-      nativeTheme.themeSource = 'light';
-    }
-    event.sender.send(IPC_DARK_MODE, nativeTheme.shouldUseDarkColors);
   });
 }
