+++ frontend/app/src/electron-main/ipc-commands.ts
@@ -1,0 +1,18 @@
+export const IPC_RESTART_BACKEND = 'RESTART_BACKEND' as const;
+export const IPC_CHECK_FOR_UPDATES = 'CHECK_FOR_UPDATES' as const;
+export const IPC_DOWNLOAD_UPDATE = 'DOWNLOAD_UPDATE' as const;
+export const IPC_DOWNLOAD_PROGRESS = 'DOWNLOAD_PROGRESS' as const;
+export const IPC_INSTALL_UPDATE = 'INSTALL_UPDATE' as const;
+export const IPC_DARK_MODE = 'DARK_MODE' as const;
+export const IPC_VERSION = 'VERSION' as const;
+export const IPC_ABOUT = 'ABOUT' as const;
+export const IPC_OPEN_PATH = 'OPEN_PATH' as const;
+export const IPC_CONFIG = 'CONFIG' as const;
+export const IPC_METAMASK_IMPORT = 'METAMASK_IMPORT' as const;
+export const IPC_SERVER_URL = 'SERVER_URL' as const;
+export const IPC_GET_DEBUG = 'GET_DEBUG' as const;
+export const IPC_OPEN_URL = 'OPEN_URL' as const;
+export const IPC_OPEN_DIRECTORY = 'OPEN_DIRECTORY' as const;
+export const IPC_DEBUG_SETTINGS = 'DEBUG_SETTINGS' as const;
+export const IPC_CLOSE_APP = 'CLOSE_APP' as const;
+export const IPC_PREMIUM_LOGIN = 'PREMIUM_USER_LOGGED_IN' as const;
